Candidate with contactDetails - list()/get()/save() works, but update() doesn't work yet.

+++ src/main/java/com/staff/controller/CandidateController.java
@@ -3,7 +3,6 @@ package com.staff.controller;
 import com.staff.dao.CandidateDao;
 import com.staff.model.Candidate;
 import com.staff.model.Candidate_;
-import com.staff.model.ContactDetails;
 import com.staff.modelDto.CandidateDto;
 import com.staff.util.filtering.CandidateFilter;
 import org.modelmapper.ModelMapper;
@@ -13,10 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -35,31 +31,14 @@ public class CandidateController {
     private final Logger logger = LoggerFactory.getLogger(CandidateController.class);
 
     @PostMapping(value = "/candidate", consumes="application/json;charset=UTF-8" )
-    public void save(@RequestBody CandidateDto candidateDto)  throws ParseException {
-
-        Candidate candidate = new Candidate();
-        Candidate candidateIn = convertToEntity(candidateDto);
-        candidate.setCandidateState( candidateIn.getCandidateState());
-        candidate.setName( candidateIn.getName());
-        candidate.setSurname( candidateIn.getSurname());
-        candidate.setSalary( candidateIn.getSalary());
-        candidate.setBirthday( candidateIn.getBirthday());
-        //candidate.setContactDetailsList(contactDetailsEmpty);
-        Long id = candidateDao.save( candidate );
-        logger.debug("saved: id = "  + id.toString());
-        List<ContactDetails> contactDetails = convertToEntity(candidateDto).getContactDetailsList();
-        List<ContactDetails> contactDetailsEmpty = new ArrayList<>();
-        ContactDetails contactDetailsLinked;
-        for ( ContactDetails contactDetailsOne : contactDetails) {
-            contactDetailsLinked = contactDetailsOne;
-            contactDetailsLinked.setCandidate(candidate);
-            contactDetailsEmpty.add(contactDetailsLinked);
+    public ResponseEntity<CandidateDto> save(@RequestBody CandidateDto candidateDto)  throws ParseException {
+        Long id = candidateDao.save( convertToEntity(candidateDto) );
+        Candidate candidate = candidateDao.get(id);
+        if(candidate == null){
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(null);
+        } else {
+            return ResponseEntity.ok().body(convertToDto(candidate));
         }
-        candidate.setContactDetailsList(contactDetailsEmpty);
-        candidateDao.update( id, candidate );
-        logger.debug("updated: " + id.toString() + " size: " + contactDetailsEmpty.size());
-        //logger.debug("done. candidate: " + candidate.toString());
-        //return ResponseEntity.ok().body("New candidate has been saved with ID:" + id);
     }
 
     @GetMapping(value = "/candidate/{id}", produces="application/json;charset=UTF-8")
@@ -67,11 +46,9 @@ public class CandidateController {
         Candidate candidate = candidateDao.get(id);
         if(candidate == null){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-        }else {
+        } else {
             return ResponseEntity.ok().body(convertToDto(candidate));
         }
-        //return convertToDto(candidate);
-        //return ResponseEntity.ok().body(candidate);
     }
 
     @GetMapping("/candidate")
@@ -101,48 +78,43 @@ public class CandidateController {
         candidateFilter.setOrder(order);
         candidateFilter.setCandidateStates(listCandidateStates);
         List<Candidate> candidates = candidateDao.list(candidateFilter);
-        logger.debug("done, recs: " + candidates.size());
-
         return candidates.stream()
                 .map(candidate -> convertToDto(candidate))
                 .collect(Collectors.toList());
-        //return candidates;
-        //return ResponseEntity.ok().body(candidates);
     }
 
     @PutMapping(value = "/candidate/{id}", consumes="application/json;charset=UTF-8")
-    public void update(@PathVariable("id") Long id, @RequestBody CandidateDto candidate) throws ParseException {
-        candidateDao.update(id, convertToEntity(candidate));
-        //return ResponseEntity.ok().body("Candidate [id=" + id.toString() + "] has been updated successfully.");
+    public ResponseEntity<CandidateDto> update(@PathVariable("id") Long id, @RequestBody CandidateDto candidateDto) throws ParseException {
+        Candidate candidate = candidateDao.get(id);
+        if(candidate == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        } else {
+            candidateDao.update(id, convertToEntity(candidateDto));
+            candidate = candidateDao.get(id);
+            return ResponseEntity.ok().body(convertToDto(candidate));
+        }
     }
 
     @DeleteMapping("/candidate/{id}")
-    public void delete(@PathVariable("id") Long id) {
-        candidateDao.delete(id);
-        //return ResponseEntity.ok().body("Candidate [id=" + id.toString() + "] has been deleted successfully.");
+    public ResponseEntity<?> delete(@PathVariable("id") Long id) {
+        Candidate candidate = candidateDao.get(id);
+        if(candidate == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        } else {
+            candidateDao.delete(id);
+            return ResponseEntity.ok().body(null);
+        }
     }
 
     private CandidateDto convertToDto(Candidate candidate) {
-        //CandidateDto candidateDto = new CandidateDto();
-        //if (candidate != null) {
         CandidateDto candidateDto = modelMapper.map(candidate, CandidateDto.class);
-        //}
-        /*for ( ContactDetails cd : candidate.getContactDetailsList()) {
-            //candidateDto.getContactDetailsList().add(cd);
-            logger.debug("convertToDto, cd: " + cd.getContactDetails()+" " +cd.getContactType());
-        }*/
-        /*postDto.setSubmissionDate(post.getSubmissionDate(),
-        userService.getCurrentUser().getPreference().getTimezone());*/
+        candidateDto.setCandidateStateAsString(candidate.getCandidateState().getDescription());
         return candidateDto;
     }
 
-    private Candidate convertToEntity(CandidateDto candidateDto) throws ParseException {
-        //Candidate candidate = new Candidate();
-        //if (candidateDto != null) {
-            Candidate candidate = modelMapper.map(candidateDto, Candidate.class);
-        //}
-        /*post.setSubmissionDate(postDto.getSubmissionDateConverted(
-                userService.getCurrentUser().getPreference().getTimezone()));*/
+    private Candidate convertToEntity(CandidateDto candidateDto) {
+        Candidate candidate = modelMapper.map(candidateDto, Candidate.class);
+        candidate.setCandidateState(candidateDto.getCandidateStateAsString());
         return candidate;
     }
 }
